function-app/requirements.txt
azure-functions
azure-functions-durable
azure-identity
azure-mgmt-containerinstance
requests
